+++ src/fobi/__init__.py
@@ -1,6 +1,6 @@
 __title__ = 'django-fobi'
-__version__ = '0.8.3'
-__build__ = 0x00005a
+__version__ = '0.8.4'
+__build__ = 0x00005b
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = '2014-2016 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'